Add files via upload
Working Sedov version supporting file reading and generation. 
Usage: replace your sedov.cpp with this one and add the SedovInputFileReader.hpp, recompile.
Not specifying a input file with the command line flag "--input" will default to the generator.
Input files as of now need to be SPH-EXA compatible binary files.